Add translations to batteries components
