+++ public/languages/en/common.json
@@ -174,6 +174,7 @@
     "boxes": {
         "acLoads": "AC Loads",
         "batteries": "Batteries",
+        "BatteriesOverview": "Batteries",
         "battery": "Battery",
         "batteryWithName": "Battery: %{name}",
         "chargerWithName": "Charger: %{productNameShort}",
