--- src/components/boxes/Batteries/index.ts
@@ -1,2 +1,0 @@
-export { default } from './Batteries'
-export * from './Batteries'
